requirements.txt
Jetson.GPIO==2.1.6
RPi.GPIO>=0.7.0